requirements.txt
amqp==2.4.2
beautifulsoup4==4.7.1
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-bootstrap4==0.0.8
django-braces==1.13.0
django-chartjs==1.3
idna==2.8
kombu==4.3.0
lxml==4.3.3
pycurl==7.43.0.2
python-dateutil==2.8.0
pytils==0.3
pytz==2018.9
requests==2.21.0
selection==0.0.14
six==1.12.0
soupsieve==1.9
urllib3==1.24.1
user-agent==0.1.9
vine==1.3.0
weblib==0.1.30
